Implement the copy credentials as prefetching
